prevent implicit routing

+++ app/controllers/CrudController.php
@@ -116,13 +116,17 @@ class CrudController extends Controller {
     }
 
     function delete($id){
-        $this->CrudModel->delete($id);
-        redirect('trash');
+        if($this->lauth->get_role($id) == "admin") {
+            $this->CrudModel->delete($id);
+            redirect('trash');
+        }
     }
 
     function soft_delete($id){
-        $this->CrudModel->soft_delete($id);
-        redirect();
+        if($this->lauth->get_role($id) == "admin") {
+            $this->CrudModel->soft_delete($id);
+            redirect();
+        }
     }
 
     function trash(){
@@ -155,7 +159,9 @@ class CrudController extends Controller {
     }
 
     function restore($id){
-        $this->CrudModel->restore($id);
-        redirect('trash');
+        if($this->lauth->get_role($id) == "admin") {
+            $this->CrudModel->restore($id);
+            redirect('trash');
+        }
     }
 }
